minor changes on product card
edited some displays, edited delete method for stores and products,
edited product card displays: do not display products from deleted stores, edited display of "per *producunit*"

+++ adminPages/deleteStore.php
@@ -16,27 +16,65 @@
 		    //get sellerID
 		    $sellerID =  filter_var($_GET['sellerID'], FILTER_SANITIZE_NUMBER_INT);
 
-		    $sql = "DELETE FROM sellers WHERE sellerID = ?";
-		    $sql = $conn->prepare($sql);
+		    /* check if the store status */
+		    $status = "SELECT storeStatus FROM sellers WHERE sellerID = ?";
+			$status = $conn->prepare($status);
+		    $status->bind_param('s', $sellerID);
 
-		    $sql->bind_param('s', $sellerID);
+		    if ($status->execute() === TRUE)
+		  	{
+		    	$result = $status->get_result();
+		    	$row = $result->fetch_assoc();
+
+		    	/* if the store/seller being deleted is a store application */
+		    	if($row['storeStatus'] == 0 )
+		    	{
+		    		$sql = "DELETE FROM sellers WHERE sellerID = ?";
+				    $sql = $conn->prepare($sql);
+				    $sql->bind_param('s', $sellerID);
+
+				    if ($sql->execute() === TRUE)
+				  	{
+						success();
+				  	}	
+				  	else
+				  	{
+				  		error();
+				  	}
+		    	}
+
+		    	/* if the store/seller is an existing store/seller */
+		    	else
+		    	{
+		    		// update seller status to -1
+			    	$sql = "UPDATE `sellers` SET storeStatus='-1' WHERE sellerID = ?";
+				    $sql = $conn->prepare($sql);
+				    $sql->bind_param('s', $sellerID);
+				    
+				    // update seller products to -1
+				    $update_products = "UPDATE `product` SET stock ='-1' WHERE seller = ?";
+				    $update_products = $conn->prepare($update_products);
+				    $update_products->bind_param('s', $sellerID);
+
+				  	if ($sql->execute() === TRUE)
+				  	{
+				    	$result = $sql->get_result();
+
+				    	if($update_products->execute() == TRUE)
+				    	{
+				    		$result = $sql->get_result();
+							success();
+				    	}
+				  	}	
+				  	else
+				  	{
+				  		error();
+				  	}
+					
+		    	}
+			}
+
 		    
-		  	if ($sql->execute() === TRUE)
-		  	{
-		    	$result = $sql->get_result();
-				echo '<script>
-						alert("Success: Store application deleted.");
-						window.location.href="admin_main.php";
-				</script>';
-		  	}	
-		  	else
-		  	{
-		  		echo die ('Error updating database<br />' . mysqli_errno($conn) . ": " . mysqli_error($conn));
-				// echo '<script>
-				// 		alert("Error: Store application was not deleted.");
-				// 		window.location.href="admin_main.php";
-				// </script>';
-		  	}
 
 		  	$conn->close();
 		}
@@ -47,5 +85,25 @@
 		}
 	}
 	else
-		header("Location: ../loginPage/login.php");
+		Header("Location: ../loginPage/login.php");
+?>
+
+<?php
+
+	function success()
+	{
+		echo '<script>
+					alert("Success: Store application deleted.");
+					window.location.href="admin_main.php";
+			</script>';
+	}
+
+	function error()
+	{
+		//echo die ('Error updating database<br />' . mysqli_errno($conn) . ": " . mysqli_error($conn));
+		echo '<script>
+				alert("Error: Store application was not deleted.");
+				window.location.href="admin_main.php";
+		</script>';
+	}	
 ?>
